refactor: renome folder name from `commands` to `command`
