+++ src/command/createNote.mts
