+++ src/command/genSettingsJson/index.mts
