+++ src/command/utils/expandSnippetPlaceholders.mts
