+++ src/command/utils/parseJsonc.mts
