+++ test/command/getSettingsJson/index.test.mts
